Fix broken nav items borders of nested Tabs (#402)
The nav items borders of a nested Tabs element are broken. They are broken because we use Descendant Combinator Selector (space) that apply the style to all the descendant elements that match a selector. So, the outer Tabs element affects the nested Tabs element style because, after all, they have common selectors. To overcome this conflict we should use Child Combinator Selector (>) to scope the style and prevent the applying the outer Tabs style to the nested Tabs element style.

Co-authored-by: Jurriaan Roelofs <jur@dxpr.com>