Fix issue with regions with only html content not being rendered
Regions with only html content such as <img> or <form> were not
displayed due to all html being stripped away. This commit fixes
that issue by using a macro to check if the region is empty. In
addition two new `has_content_*` variables are used in the logic
checks to prevent multiple renders of the sidebars.

modified: templates/page.html.twig