Fix issue with missing third-party block content class
This commit fixes an issue where a block's 'Content CSS class(es)'
field value was never applied due to the block template not having
a wrapper class fro the block content. Instead this class is now
applied to the content's wrapper.

A change was also done to the main menu template to avoid that the
content class is being applied to every level of the menu, and
instead only being applied to the top level, menu level 0.

modified: dxpr_theme.theme
modified: templates/menu--main.html.twig