Refactor admin settings page title and add no-preview info
This commits refactors the admin settings page title which was all
added to a form element prefix key. The prefix key was removed and
a new "inline_template" form element was added to display the page
title instead. This was mainly done as a clean-up when adding a no
live-preview icon explanation was added.

modified: theme-settings.php