Fix for images larger than container display area
When an image width is wider than the container the width is
adjusted to fit the container, but the height stays the same,
effectively squeezing images together. This commits sets the
default height for all images to 'auto' so that the aspect ratio
is retained when image width is changed.

modified: css/base/layout.css
modified: scss/base/layout.scss